+++ app/helpers/application_helper.rb
@@ -51,19 +51,17 @@ module ApplicationHelper
     end.html_safe
   end
 
-  # TODO: Update (lowercase)
-  def th_sort_field(order, sort_field, display_name)
+  def th_sort_field(order, sort_field, display_name, extra_class: '')
     sort_params = params.permit(:search)
-
-    sort_field_order = (order == sort_field) ? "#{sort_field} DESC" : sort_field
+    sort_field_order = (order == sort_field) ? "#{sort_field} desc" : sort_field
     if order == sort_field
       css_class = 'sort-up'
       selected_class = 'sort-selected'
-    elsif order == sort_field + ' DESC NULLS LAST'
+    elsif order == sort_field + ' desc nulls last'
       css_class = 'sort-down'
       selected_class = 'sort-selected'
     end
-    content_tag(:th, class: ['nowrap', selected_class]) do
+    content_tag(:th, class: [selected_class, extra_class]) do
       link_to url_for(sort_params.merge(order: sort_field_order)), style: 'text-decoration:none', class: css_class do
         display_name.to_s.html_safe
       end