+++ app/models/broadcast.rb
@@ -4,7 +4,10 @@
 # set to be published on specific dates.
 class Broadcast < ApplicationRecord
   # Concerns
-  include Deletable
+  include Deletable, Replyable
+  include PgSearch
+  multisearchable against: [:title, :short_description, :keywords, :description],
+                  unless: :deleted?
 
   # Named Scopes
   scope :published, -> { current.where(published: true).where('publish_date <= ?', Time.zone.today) }
@@ -17,13 +20,16 @@ class Broadcast < ApplicationRecord
   # Model Relationships
   belongs_to :user
   belongs_to :category
-  has_many :broadcast_comments
-  has_many :broadcast_comment_users
 
   # Model Methods
+  def destroy
+    super
+    update_pg_search_document
+    replies.each(&:update_pg_search_document)
+  end
 
   def to_param
-    slug.to_s
+    slug_was.to_s
   end
 
   def url_hash
@@ -36,5 +42,42 @@ class Broadcast < ApplicationRecord
 
   def editable_by?(current_user)
     current_user.editable_broadcasts.where(id: id).count == 1
+  end
+
+  def last_page
+    # ((replies.where(reply_id: nil).count - 1) / Reply::REPLIES_PER_PAGE) + 1
+    1
+  end
+
+
+  # TODO: Refactor or remove into search module
+  def self.compute_ranges(description_array, terms)
+    cleaned_array = description_array.collect { |t| t.gsub(/[^\w]/, '').to_s.downcase }
+    indices = cleaned_array.each_index.select { |i| cleaned_array[i].in?(terms) }
+    @ranges = []
+    indices.each do |index|
+      @ranges << [[index - 4, 0].max, [index + 4, cleaned_array.count].min]
+    end
+    @ranges = merge_overlapping_ranges(@ranges)
+    @ranges = [[0, 9]] if @ranges.empty?
+    @ranges
+  end
+
+  def self.ranges_overlap?(a, b)
+    a.last >= b.first
+  end
+
+  def self.merge_ranges(a, b)
+    [[a.first, b.first].min, [a.last, b.last].max]
+  end
+
+  def self.merge_overlapping_ranges(ranges)
+    ranges.sort_by(&:first).inject([]) do |inner_ranges, range|
+      if !inner_ranges.empty? && ranges_overlap?(inner_ranges.last, range)
+        inner_ranges[0...-1] + [merge_ranges(inner_ranges.last, range)]
+      else
+        inner_ranges + [range]
+      end
+    end
   end
 end