+++ db/migrate/20160620163138_add_view_count_to_broadcasts.rb
@@ -1,0 +1,6 @@
+class AddViewCountToBroadcasts < ActiveRecord::Migration
+  def change
+    add_column :broadcasts, :view_count, :integer, null: false, default: 0
+    add_index :broadcasts, :view_count
+  end
+end