+++ test/controllers/admin_controller_test.rb
@@ -88,4 +88,26 @@ class AdminControllerTest < ActionController::TestCase
     get :sync
     assert_redirected_to new_user_session_path
   end
+
+  test 'should get agreement reports for system admin' do
+    login(users(:admin))
+    get :agreement_reports
+    assert_response :success
+  end
+
+  test 'should not get agreement reports for non system admin' do
+    get :agreement_reports
+    assert_redirected_to new_user_session_path
+  end
+
+  test 'should get downloads by month for system admin' do
+    login(users(:admin))
+    get :downloads_by_month
+    assert_response :success
+  end
+
+  test 'should not get downloads by month for non system admin' do
+    get :downloads_by_month
+    assert_redirected_to new_user_session_path
+  end
 end