+++ test/controllers/replies_controller_test.rb
@@ -32,7 +32,7 @@ class RepliesControllerTest < ActionController::TestCase
   test 'should preview reply' do
     login(@regular_user)
     post :preview, params: {
-      parent_comment_id: 'root', reply_id: 'new', reply: reply_params
+      parent_reply_id: 'root', reply_id: 'new', reply: reply_params
     }, format: 'js'
     assert_template 'preview'
     assert_response :success
@@ -49,13 +49,11 @@ class RepliesControllerTest < ActionController::TestCase
     assert_response :success
   end
 
-  # TODO: Should add test to reply to blog post.
   test 'should create reply to blog post' do
-    skip
     login(@regular_user)
     assert_difference('Reply.count') do
       post :create, params: {
-        broadcast_id: broadcasts(:one).to_param, reply: reply_params
+        broadcast_id: broadcasts(:published).to_param, reply: reply_params
       }, format: 'js'
     end
     assert_template 'create'
@@ -68,12 +66,10 @@ class RepliesControllerTest < ActionController::TestCase
     assert_redirected_to page_topic_path(@reply.topic, page: @reply.page, anchor: @reply.anchor)
   end
 
-  # TODO: Add a redirect for a reply on a blog post.
   test 'should show reply to a blog post and redirect to correct page' do
-    skip
     login(@regular_user)
-    get :show, params: { id: @reply_blog_post }
-    assert_redirected_to page_broadcast_path(@reply_blog_post.broadcast, page: @reply_blog_post.page, anchor: @reply_blog_post.anchor)
+    get :show, params: { id: replies(:blog_one) }
+    assert_redirected_to blog_post_path(broadcasts(:published).url_hash.merge(page: replies(:blog_one).page, anchor: replies(:blog_one).anchor))
   end
 
   test 'should show reply' do